Move hasIssuerName
